pyproject.toml
[project]
name = "thesis-gateway"
version = "0.1.0"
description = "This is the thesis API gateway"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "AxCodes" }
]
dependencies = [
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]

[tool.pytest.ini_options]
pythonpath = "."

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]  # E: pycodestyle errors, F: pyflakes, I: isort (import sorting)
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]